Modify:피드 댓글 공유 오류 수정
